fix if command run from console

+++ src/MonitorsSchedule.php
@@ -19,45 +19,47 @@ trait MonitorsSchedule
 	 */
 	public function monitor(Schedule $schedule)
 	{
+		if ( !$this->app->runningInConsole() ) {
+			$date = Carbon::today()->toDateString();
 
-		$date = Carbon::today()->toDateString();
+			$events = new Collection($schedule->events());
 
-		$events = new Collection($schedule->events());
+			$events->each(function (Event $event) use ($date) {
 
-		$events->each(function (Event $event) use ($date) {
+				$command = str_after($event->command, '\'artisan\' ');
+				$filename = str_slug($command) . "-$date.log";
+				$path = storage_path("logs/$filename");
 
-			$command = str_after($event->command, '\'artisan\' ');
-			$filename = str_slug($command) . "-$date.log";
-			$path = storage_path("logs/$filename");
+				$event->sendOutputTo($path)->after(function () use ($command, $path) {
 
-			$event->sendOutputTo($path)->after(function () use ($command, $path) {
+					if (file_exists($path) && ($output = file_get_contents($path))) {
 
-				if (file_exists($path) && ($output = file_get_contents($path))) {
+						$start = $end = NULL;
+						$json = json_decode($output, true);
 
-					$start = $end = NULL;
-					$json = json_decode($output, true);
+						if( !empty( $json ) ){
+							// cron success
+							$start = $json['start'];
+							$end = $json['end'];
+							$output = $json['output'];
+							Cron::where( 'command', $command )->increment( 'success', 1);
+						} else {
+							Cron::where( 'command', $command )->increment( 'error', 1);
+						}
 
-					if( !empty( $json ) ){
-						// cron success
-						$start = $json['start'];
-						$end = $json['end'];
-						$output = $json['output'];
-						Cron::where( 'command', $command )->increment( 'success', 1);
-					} else {
-						Cron::where( 'command', $command )->increment( 'error', 1);
+						DB::table('cron_loggers')->insert([
+							'command'   => $command,
+							'start'     => $start,
+							'end'		=> $end,
+							'output'    => $output,
+							'created_at' => Carbon::now(),
+						]);
+
+						unlink($path);
 					}
-
-					DB::table('cron_loggers')->insert([
-						'command'   => $command,
-						'start'     => $start,
-						'end'		=> $end,
-						'output'    => $output,
-						'created_at' => Carbon::now(),
-					]);
-
-					unlink($path);
-				}
+				});
 			});
-		});
+		}
+
 	}
 }
